move s3 read function to Action class so it can be used by other actions
